refactor: modernize LoginScreen source and fix validator bugs
- Move source from lib/ to src/.
- Fix passwordValidator regex (\\d -> \d) so the digit rule actually applies.
- Guard LayoutAnimation on Android via setLayoutAnimationEnabledExperimental.
- Replace any-typed props (logoImageSource, children, TouchableComponent) and
  rewrite useStateWithCallback with real generics.
- Wrap LoginScreen in React.memo; memoize render helpers and tooltip content.
- Derive widths from useWindowDimensions instead of a module-load Dimensions read.
- Add accessibilityRole/accessibilityLabel to buttons and inputs.
- Export prop types and password pattern presets from the barrel.
