+++ src/index.ts
@@ -5,3 +5,14 @@ import passwordValidator from "./helpers/passwordValidator";
 
 export default LoginScreen;
 export { SocialButton, emailValidator, passwordValidator };
+export {
+  patternNormal,
+  patternMedium,
+  patternHigh,
+} from "./helpers/passwordValidator";
+
+export type { ILoginScreenProps } from "./LoginScreen";
+export type {
+  ISocialButtonProps,
+  TouchableComponentType,
+} from "./components/social-button/SocialButton";
