+++ tsconfig.json
@@ -1,39 +1,27 @@
 {
   "compilerOptions": {
     "target": "esnext",
-    "module": "es6",
-    "lib": ["es6"],
-    "allowJs": true,
+    "module": "esnext",
+    "lib": ["esnext", "dom"],
     "jsx": "react-native",
-    "noImplicitAny": false,
-    "incremental": true /* Enable incremental compilation */,
-    "isolatedModules": true,
-    "strict": true,
     "moduleResolution": "node",
     "baseUrl": "./",
-    "outDir": "build/dist",
-    "noEmitHelpers": true,
+    "rootDir": "src",
+    "allowJs": false,
+    "declaration": true,
+    "isolatedModules": true,
+    "strict": true,
+    "noImplicitAny": true,
     "alwaysStrict": true,
     "strictFunctionTypes": true,
-    "resolveJsonModule": true,
-    "importHelpers": false,
-    "experimentalDecorators": true,
-    "strictPropertyInitialization": false,
-    "allowSyntheticDefaultImports": true,
     "strictNullChecks": true,
-    "skipDefaultLibCheck": true,
-    "skipLibCheck": true,
+    "strictPropertyInitialization": false,
+    "resolveJsonModule": true,
+    "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
-    "typeRoots": ["./node_modules/@types", "./@types"],
-    "declaration": true /* Generates corresponding '.d.ts' file. */,
-    "sourceMap": true /* Generates corresponding '.map' file. */
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true
   },
-  "exclude": [
-    "example",
-    "example-manual-state",
-    "node_modules",
-    "babel.config.js",
-    "metro.config.js",
-    "jest.config.js"
-  ]
+  "include": ["src"],
+  "exclude": ["node_modules", "lib", "build", "example"]
 }
